Move property_set and property_set_jsonb_dictionary_builder to properties module
Better organization - these modules logically belong with other property-related functionality

- Moved property_set.rs to properties/property_set.rs
- Moved property_set_jsonb_dictionary_builder.rs to properties/property_set_jsonb_dictionary_builder.rs
- Updated module declarations in lib.rs and properties/mod.rs
- Updated imports in all dependent files
