+++ rust/analytics/src/lib.rs
@@ -39,10 +39,6 @@ pub mod metrics_table;
 pub mod payload;
 /// UDF to convert properties to dictionary-encoded arrays
 pub mod properties;
-/// Reference-counted set of properties in transit format
-pub mod property_set;
-/// Custom dictionary builder for PropertySet with pointer-based deduplication
-pub mod property_set_jsonb_dictionary_builder;
 pub mod record_batch_transformer;
 /// bulk metadata & payload ingestion using Arrow
 pub mod replication;
