+++ rust/analytics/src/properties/mod.rs
@@ -11,4 +11,6 @@ pub mod properties_column_accessor;
 pub mod properties_to_dict_udf;
 pub mod properties_to_jsonb_udf;
 pub mod property_get;
+pub mod property_set;
+pub mod property_set_jsonb_dictionary_builder;
 pub mod utils;
